attempting to describe bounce off
(I tried to push this a second ago but don't see it. Doing it again.)

See what you think of this mess.
I set what I think off by stars in the place I think it should live.

We'd need to call to it in drawPlanet, if it's a workable idea.
But tell me why it's not, Please!