Attempting to turn dummy code into real code
This isn't working, which doesn't surprise me. But what does surprise
me is the error. It's not recognizing 'x0' though it's defined a couple
of lines of code above.

I wrote it originally and then got a ton of errors. I fixed (or tried)
them but was brought down by this. It feels a little like robbing a
bank guns a blazing but then tripping on your untied shoes on the way
out.

The changes are set off by stars but I also call the new function in
drawPlanet.

By the way, I still can't quite tell if I'm committing these to the
right place or what.

@hcwiley @juggleware @RubyVerbena @kaben @br1anabb0tt @Sveltlanamonsoon
@misui8 @fguevara @markstrella @aareali @azuroma @jpatricknola
@jwgoedert @davidrhoden